refactor: 디렉토리 리팩토링 #15
