+++ src/main/java/com/testcar/car/domains/trackReservation/TrackReservationService.java
@@ -6,6 +6,8 @@ import com.testcar.car.common.exception.NotFoundException;
 import com.testcar.car.domains.member.Member;
 import com.testcar.car.domains.track.Track;
 import com.testcar.car.domains.track.TrackService;
+import com.testcar.car.domains.trackReservation.entity.ReservationStatus;
+import com.testcar.car.domains.trackReservation.entity.TrackReservation;
 import com.testcar.car.domains.trackReservation.model.TrackReservationRequest;
 import com.testcar.car.domains.trackReservation.model.vo.TrackReservationFilterCondition;
 import com.testcar.car.domains.trackReservation.repository.TrackReservationRepository;
