+++ src/main/java/com/testcar/car/domains/trackReservation/model/TrackReservationDetailResponse.java
@@ -2,9 +2,9 @@ package com.testcar.car.domains.trackReservation.model;
 
 
 import com.testcar.car.domains.member.Member;
-import com.testcar.car.domains.trackReservation.ReservationStatus;
-import com.testcar.car.domains.trackReservation.TrackReservation;
-import com.testcar.car.domains.trackReservation.TrackReservationSlot;
+import com.testcar.car.domains.trackReservation.entity.ReservationStatus;
+import com.testcar.car.domains.trackReservation.entity.TrackReservation;
+import com.testcar.car.domains.trackReservation.entity.TrackReservationSlot;
 import com.testcar.car.domains.trackReservation.model.vo.ReservationSlotVo;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
