+++ src/main/java/com/testcar/car/domains/trackReservation/repository/TrackReservationCustomRepository.java
@@ -1,7 +1,7 @@
 package com.testcar.car.domains.trackReservation.repository;
 
 
-import com.testcar.car.domains.trackReservation.TrackReservation;
+import com.testcar.car.domains.trackReservation.entity.TrackReservation;
 import com.testcar.car.domains.trackReservation.model.vo.TrackReservationFilterCondition;
 import java.util.List;
 import java.util.Optional;
