+++ src/main/java/com/testcar/car/domains/trackReservation/repository/TrackReservationCustomRepositoryImpl.java
@@ -8,8 +8,8 @@ import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.testcar.car.common.entity.BaseQueryDslRepository;
-import com.testcar.car.domains.trackReservation.ReservationStatus;
-import com.testcar.car.domains.trackReservation.TrackReservation;
+import com.testcar.car.domains.trackReservation.entity.ReservationStatus;
+import com.testcar.car.domains.trackReservation.entity.TrackReservation;
 import com.testcar.car.domains.trackReservation.model.vo.TrackReservationFilterCondition;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
