+++ src/main/java/com/testcar/car/domains/trackReservation/repository/TrackReservationRepository.java
@@ -1,7 +1,7 @@
 package com.testcar.car.domains.trackReservation.repository;
 
 
-import com.testcar.car.domains.trackReservation.TrackReservation;
+import com.testcar.car.domains.trackReservation.entity.TrackReservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TrackReservationRepository
