+++ src/main/java/com/testcar/car/domains/trackReservation/repository/TrackReservationSlotCustomRepository.java
@@ -1,7 +1,7 @@
 package com.testcar.car.domains.trackReservation.repository;
 
 
-import com.testcar.car.domains.trackReservation.TrackReservationSlot;
+import com.testcar.car.domains.trackReservation.entity.TrackReservationSlot;
 import com.testcar.car.domains.trackReservation.model.vo.ReservationSlotVo;
 import java.time.LocalDate;
 import java.util.List;
